+++ src/commands/index.ts
@@ -8,6 +8,7 @@ import { join } from "./join.js";
 import { basic } from "./basic.js";
 import { oAuth2 } from "./oAuth2.js";
 import { preferences } from "./preferences.js";
+import { inventory_adjustment } from "./inventory-adjustment.js";
 
 /**
  * Route Command Event - Function
@@ -33,6 +34,8 @@ export const commands = async (CommandEvent: CommandEvent) => {
         return await join(CommandEvent);
       case "oAuth2":
         return await oAuth2(CommandEvent);
+      case "inventory_adjustment":
+        return await inventory_adjustment(CommandEvent);
       default:
         throw `Route: ${CommandEvent.command} not found`;
     }
@@ -74,5 +77,11 @@ export const commandsList: Command[] = [
     name: "Sync taxes",
     description:
       "this command to syncing all quickbooks items with your repzo taxes",
+  },
+  {
+    command: "inventory_adjustment",
+    name: "Inventory Adjustment",
+    description:
+      "this command shall create an inventory adjustment with the difference between all accumulated repzo warehouses and quickbooks inventory making quickbooks similar to Repzo accumulated inventory",
   },
 ];